Changed names to index and add checking for current user

+++ src/iTechArt.SurveysSite.WebApp/Controllers/SurveyController.cs
@@ -25,7 +25,7 @@ namespace iTechArt.SurveysSite.WebApp.Controllers
 
 
         [HttpGet]
-        public async Task<IActionResult> DisplayAll()
+        public async Task<IActionResult> Index()
         {
             var userId = User.GetId();
             var surveys = await _surveyService.GetAllUserSurveysAsync(userId);
@@ -75,15 +75,29 @@ namespace iTechArt.SurveysSite.WebApp.Controllers
         public async Task<IActionResult> Delete(int id)
         {
             var survey = await _surveyService.GetByIdAsync(id);
+            var userId = User.GetId();
+
+            if (userId != survey.Owner.Id)
+            {
+                return RedirectToAction("AccessDenied", "Home");
+            }
+
             await _surveyService.DeleteSurveyAsync(survey);
 
-            return RedirectToAction("DisplayAll");
+            return RedirectToAction("Index");
         }
 
         [HttpGet]
         public async Task<IActionResult> Edit(int id)
         {
             var survey = await _surveyService.GetByIdAsync(id);
+            var userId = User.GetId();
+
+            if (userId != survey.Owner.Id)
+            {
+                return RedirectToAction("AccessDenied", "Home");
+            }
+
             var surveyViewModel = new SurveyViewModel
             {
                 Id = survey.Id,
@@ -102,16 +116,9 @@ namespace iTechArt.SurveysSite.WebApp.Controllers
                 return View(surveyViewModel);
             }
 
-            var userId = User.GetId();
-            var user = await _userManagementService.GetUserByIdAsync(userId);
-
-            var survey = new Survey
-            {
-                Id = surveyViewModel.Id,
-                ChangeDate = DateTime.Now,
-                Owner = user,
-                Title = surveyViewModel.Title
-            };
+            var survey = await _surveyService.GetByIdAsync(surveyViewModel.Id);
+            survey.Title = surveyViewModel.Title;
+            survey.ChangeDate = DateTime.Now;
 
             await _surveyService.UpdateSurveyAsync(survey);
